third_party/blink/web_tests/fast/webmcp/execute_tool.html

<!DOCTYPE html>
<html>
<head>
<form toolautosubmit toolname="search_tool" tooldescription="Search the web" target="_blank">
  <input type="text" name="query">
</form>
<script>
async function echo(obj) {
  return obj.text;
}

navigator.modelContext.registerTool({
  execute: echo,
  name: "echo",
  description: "echo input",
  inputSchema: {
      type: "object",
      properties: {
          "text": {
              description: "Value to echo",
              type: "string",
          }
      },
      required: ["text"]
  },
});

async function longRunning() {
  await new Promise(r => { setTimeout(r, 5000); });
}

navigator.modelContext.registerTool({
  execute: longRunning,
  name: "longRunning",
  description: "show tool",
  inputSchema: {
      type: "object",
      properties: {}
  },
});

function print(message) {
    document.getElementById("console").appendChild(document.createTextNode(message + "\n"));
}

async function runTests() {
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }

    if (!window.navigator.modelContext) {
        print("FAIL: modelContext API disabled");
        return;
    }

    if (!window.navigator.modelContextTesting) {
        print("FAIL: Testing API disabled");
        return;
    }

    // Correct same-document case
    try {
        const inputArgs = JSON.stringify({ text: "Hello world" });
        const result = await navigator.modelContextTesting.executeTool("echo", inputArgs);

        if (result === "Hello world") {
            print("PASS");
        } else {
            print(`FAIL: Expected 'Hello world' but got '${result}'`);
        }
    } catch (error) {
        print(`FAIL: ${error.message}`);
    }

    // Correct cross-document case
    try {
        const inputArgs = JSON.stringify({ query: "testing" });
        const result = await navigator.modelContextTesting.executeTool("search_tool", inputArgs);

        if (result === null) {
            print("PASS");
        } else {
            print(`FAIL: Expected null but got '${result}'`);
        }
    } catch (error) {
        print(`FAIL: ${error.message}`);
    }

    // Test case for a non-existent tool
    try {
        const result = await navigator.modelContextTesting.executeTool("nonExistentTool", "{}");
        print(`FAIL: executeTool should have thrown an error for a non-existent tool but got '${result}'`);
    } catch (error) {
        print(`INFO: Received expected error for non-existent tool: ${error.name}`);
    }

    // Test case for invalid JSON input
    try {
        const result = await navigator.modelContextTesting.executeTool("echo", "{invalid_json: 'test'");
        print(`FAIL: executeTool should have thrown an error for invalid JSON input but got '${result}'`);
    } catch (error) {
        print(`INFO: Received expected error for invalid JSON: ${error.name}`);
    }

    // Test case for input that doesn't match schema
    try {
        const result = await navigator.modelContextTesting.executeTool("echo", JSON.stringify({ wrong_property: "test" }));
        // TODO(khushalsagar): Add input schema validation, see crbug.com/454460142.
        print(`FAIL: executeTool should have thrown an error for input mismatching schema but got '${result}'`);
    } catch (error) {
        print(`INFO: Received expected error for schema mismatch: ${error.name}`);
    }

    // Test cases for cancelling tools.
    const controller = new AbortController();
    const signal = controller.signal;
    try {
        setTimeout(() => { controller.abort('aborted after 500ms (this is expected)'); }, 500);
        const result = await navigator.modelContextTesting.executeTool("longRunning", "{}", { signal });
        print(`FAIL: executeTool should have thrown an error after 500ms but got '${result}'`);
    } catch (error) {
        print(`INFO: Received expected error for aborting after 500ms: ${error.name}`);
    }
    try {
        const result = await navigator.modelContextTesting.executeTool('longRunning', '{}', { signal });
        print(`FAIL: executeTool should have thrown an error right away but got '${result}'`);
    } catch (error) {
        print(`INFO: Received expected error for aborted signal: ${error.name}`);
    }

    if (window.testRunner) {
        testRunner.notifyDone();
    }
}

</script>
</head>
<body onload="runTests();">
This tests that document.navigator.modelContextTesting works correctly.
<pre id="console">
</pre>
</body>
</html>
